feat: change transaction to use aop module
