+++ src/modules/auth/auth.service.ts
@@ -20,9 +20,9 @@ import { Auth } from './entities/auth.entity';
 import { CashKeys, GENERAL_CACHE } from '../cache/cache.constant';
 import { CacheService } from '../cache/cache.service';
 import { ConfigsService } from '../configs/configs.service';
-import { Transactional } from '../database/transactional.decorator';
 import { MailQueueOps } from '../mail/mail.constant';
 import { InjectMailQueue } from '../queue/queue.decorator';
+import { Transactional } from '../transaction/transaciton.decorator';
 import { User } from '../users/entities/user.entity';
 import { UserRole } from '../users/users.interface';
 import { UsersService } from '../users/users.service';
