+++ src/modules/transaction/transaciton.decorator.ts
@@ -1,0 +1,5 @@
+import { TRANSACTION_DECORATOR } from './transaction.constant';
+import { TransactionalOptions } from './transaction.interface';
+import { createDecorator } from '../aop/aop.util';
+
+export const Transactional = (options?: TransactionalOptions) => createDecorator(TRANSACTION_DECORATOR, options);
