+++ src/modules/transaction/transaction.constant.ts
@@ -5,3 +5,5 @@ export const TRANSACTIONAL_OPTION = Symbol('TRANSACTIONAL_OPTION');
 // export const TYPEORM_DATA_SOURCE_NAME = '@transactional/data-source';
 //
 // export const TYPEORM_DATA_SOURCE_NAME_PREFIX = TYPEORM_DATA_SOURCE_NAME + '/';
+
+export const TRANSACTION_DECORATOR = Symbol('TransactionDecorator');
