+++ src/modules/transaction/transaction.module.ts
@@ -1,0 +1,8 @@
+import { Global, Module } from '@nestjs/common';
+import { TransactionService } from './transaction.service';
+
+@Global()
+@Module({
+  providers: [TransactionService],
+})
+export class TransactionModule {}
